+++ src/backend/resolvers/mutations/relationalResolvers/index.js
@@ -1,0 +1,5 @@
+const pathwaysAndPersonConnectionResolvers = require('./pathwaysAndPerson')
+
+module.exports = {
+  ...pathwaysAndPersonConnectionResolvers,
+}